requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
boto3==1.16.35
botocore==1.19.35
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.4
django-bootstrap4==2.3.1
django-stdimage==5.1.1
django-storages==1.10.1
gunicorn==20.0.4
importlib-metadata==2.1.1
jmespath==0.10.0
mysql==0.0.2
mysqlclient==2.0.2
Pillow==8.0.1
progressbar2==3.53.1
psycopg2-binary==2.8.6
PyMySQL==0.10.1
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.2
whitenoise==5.2.0
zipp==3.4.0
